+++ examples/router/internal/layout/model.go
@@ -1,0 +1,8 @@
+package layout
+
+import gf "github.com/graybuton/goframe/pkg/goframe"
+
+type ShellProps struct {
+	OnOpenFirstIssue func()
+	Children         []gf.Node
+}